% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atak_drapieznika.R
\name{atak_drapieznika}
\alias{atak_drapieznika}
\title{Zmniejszenie ilosci zwierzat w stadzie gracza w wyniku ataku wilka lub lisa.}
\usage{
atak_drapieznika(drapieznik, stan_stada_gracza,
  wilk_co = SuperFarmerAGOJZB::param$wilk_co,
  wilk_ile = SuperFarmerAGOJZB::param$wilk_ile,
  lis_co = SuperFarmerAGOJZB::param$lis_co,
  lis_ile = SuperFarmerAGOJZB::param$lis_ile)
}
\arguments{
\item{drapieznik}{Jest wylosowanym z kostek drapieznikiem.}

\item{stan_stada_gracza}{Aktualny stan liczbowy zwierzat w stadzie gracza.}

\item{wilk_co}{Argument jakie zwierzeta zjada wilk w postaci wektora z nazwami, domyslnie
zwierzeta z wersji podstawowej gry.}

\item{wilk_ile}{Dane o tym, ile zwierzat wilk, domyslnie dodane wartości z podstawowej
wersji gry Super Farmer.}

\item{lis_co}{Argument w postaci wektora jakie zwierzeta zjada lis, domyslnie jedynie kroliki.}

\item{lis_ile}{Dane o tym, ile krolikow zjada lis, domyslnie dodane wartości z podstawowej
wersji gry Super Farmer.}
}
\value{
Funkcja \code{atak_drapieznika} wraca zmieniony stan stada gracza.
}
\description{
Funckja \code{atak_drapieznika} przyjmujac jako argumenty nazwe drapieznika oraz
aktualny stan stada gracza, zmniejsza ilość odpowiednich zwierząt w stadzie gracza.
}
\author{
Alicja Gosiewska, \email{alicjagosiewska@gmail.com}

Joanna Zbijewska, \email{asia.zbijewska@gmail.com}

Hanna Kranas, \email{unchallenged.yo@gmail.com}
}

