% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/badaj_gre_compiled.R
\name{badaj_gre_compiled}
\alias{badaj_gre_compiled}
\title{Funkcja badajaca strategie i umozliwiajaca porownanie.}
\usage{
badaj_gre_compiled(strategia1 = SuperFarmerAGOJZB::strategia_AGOJZB,
  strategia2 = NULL, co_chcesz = "statystyki", ilosc_wywolan = 10000,
  param = SuperFarmerAGOJZB::param, sprawdzanie = TRUE)
}
\arguments{
\item{strategia1}{Argument okreslajacy pierwsza ze strategii do porównania,
ktorego domyslna wartosc to strategia zawarta w pakiecie.}

\item{strategia2}{Argument okreslajacy druga ze strategii do porownania,
domyslna wartosc przypisana to NULL.}

\item{co_chcesz}{Parametry okreslajacy rodzaj wizualizacji, który chcemy uzyskać.
Domyslna wartosc parametru to "statystyki", można użyć również parametru "histogram".}

\item{ilosc_wywolan}{Jest argumentem okreslajacym dla ilu wywolan gry chcemy otrzymac
wynik funkcji. Domyslna wartosc to 10000.}

\item{param}{Parametry okreslaja niezbedne wartosci zdefiniowane do gry.}

\item{sprawdzanie}{Argument decydujacy, czy sprawdzenie poprawnosci strategi ma byc wykonane.
Domyslnie jest wykonwane.}
}
\value{
Funkcja \code{badaj_gre} zwraca podstawowe statystyki
lub wykres(w zaleznosci od podanych parametrow).
}
\description{
Funcja \code{badaj_gre_compiled} jest skompilowaną za pomocą pakietu \code{compiler} funkcją \code{badaj_gre_compiled} wywoluje gre 10000 razy na wybranej strategii.
}
\author{
Alicja Gosiewska, \email{alicjagosiewska@gmail.com}
}

