% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gra.R
\name{gra}
\alias{gra}
\title{Przeprowadzenie gry w Super Farmera.}
\usage{
gra(strategia = SuperFarmerAGOJZB::strategia_AGOJZB,
  param = SuperFarmerAGOJZB::param, sprawdzanie = TRUE)
}
\arguments{
\item{strategia}{Argument okreslajacy wykorzystywna do gry strategie, ktorego domyslna
wartosc to strategia zawarta w pakiecie.}

\item{param}{Parametry okreslaja niezbedne wartosci zdefiniowane do gry.}

\item{sprawdzanie}{Argument decydujacy, czy sprawdzenie poprawnosci strategi ma byc wykonane.
Domyslnie jest wykonwane.}
}
\value{
Zwraca ilosc ruchow, jakie gracz musial wykonac przed zakonczeniem gry
w Super Farmera.
}
\description{
Funkcja \code{gra} przeprowadza gre dla zadanych argumentow - strategii,
jako domyslna przypisana jest strategia z pakietu, a takze parametrow gry, domyslnie
dla parametrow opisanych w oryginalnej grze Super Farmer, wyliczana jest takze liczba
ruchow jakiej gracz potrzebowal do wygranej.
}
\author{
Alicja Gosiewska, \email{alicjagosiewska@gmail.com}

Joanna Zbijewska, \email{asia.zbijewska@gmail.com}
}

