% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kilka_na_jedno.R
\name{kilka_na_jedno}
\alias{kilka_na_jedno}
\title{Wymiana kilku mniejszych zwierzat na jedno wieksze.}
\usage{
kilka_na_jedno(tabela_co, na_co, stado = SuperFarmerAGOJZB::param$stado,
  param = SuperFarmerAGOJZB::param)
}
\arguments{
\item{tabela_co}{Wektor zawierajacy zwirzeta, ktore jestesmy w stanie oddac.}

\item{na_co}{Zwierze, ktore chcemy dostac.}

\item{stado}{Zadana maksymalna ilosc zwierzat kazdego rodzaju.}

\item{param}{Przekazane parametry gry.}
}
\value{
Zwraca zmieniony stan stada gracza.
}
\description{
Funkcja \code{kilka_na_jedno} sluzy do wymiany mozliwych zwierzat ze stada gracza
na wieksze zwierzeta.
}
\author{
Alicja Gosiewska, \email{alicjagosiewska@gmail.com}
}

