% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/losuj_rzut_kostek.R
\name{losuj_rzut_kostek}
\alias{losuj_rzut_kostek}
\title{Rzut kostkami}
\usage{
losuj_rzut_kostek(z = SuperFarmerAGOJZB::param$kostka_zielona,
  c = SuperFarmerAGOJZB::param$kostka_czerwona)
}
\arguments{
\item{z}{Pierwsza kostka, z ktorej losujemy wartosc. Domyslnie to kostka zielona
odpowiadajaca kostce z zestawu do gry.}

\item{c}{Druga kostka, z ktorej losujemy wartosc. Domyslnie to kostka czerwona
odpowiadajaca tej z zestwau do gry.}
}
\value{
Funkcja \code{losuj_rzut_kostek} zwraca jako wynik wektor dwuelementowy z
wynikiem rzutu z obu kostek.
}
\description{
Funkcja \code{losuj_rzut_kostek} losuje i zwraca wynik rzutu z dwoch wskazanych kostek,
domyslnie sa to kostki zielona i czerwona, odpowiadajace kostkom z klasycznego
zestawu do gry.
}
\author{
Joanna Zbijewska, \email{asia.zbijewska@gmail.com}
}

