% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\docType{data}
\name{param}
\alias{param}
\title{Domyslne parametry gry.}
\description{
Zbior danych zwierajacy liste z domyslnymi parametrami gry. Lista zawiera:
}
\details{
stado - wektor zawierajacy maksymalna ilosc zwierzat

stado_gracza - wektor zawierajacy stan stada gracza na poczatku gry

warunek_wygranej - wektor reperzenujacy minimalna ilosc i rodzaje zwierzat potrzebne do wygranej

wartosc_zwierzat - wektor zawierajacy wartosci kazdego zwierzecia wyrazone w krolikach

kostka_zielona - wektor zawierajacy zwierzeta znajdujace sie na kazdej sciance zielonej kostki

kostka_czerwona - wektor zawierajacy zwierzeta znajdujace sie na kazdej sciance czerwonej kostki

lis_co - wektor z nazwami zwierzat, ktore zjada lis

lis_ile - wektor z iloscia zwierzat odpowiednich rodzajow pozostajacych w stadzie gracza po ataku lisa

wilk_co - wektor zawierajacy nazwy zwierzat, ktore zjada wilk podczas ataku

wilk_ile - wektor zawierajacy wartosci stan liczbowy zwierzat odpowiednich rodzajow, ktore pozostaja
w stadzie gracza po ataku wilka
}

