% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rozmnoz_stado.R
\name{rozmnoz_stado}
\alias{rozmnoz_stado}
\title{Powiekszenie ilosci zwierzat w stadzie gracza o wynik rzutu.}
\usage{
rozmnoz_stado(wynik, ilosc_kostek, stan_stada_gracza,
  stado = SuperFarmerAGOJZB::param$stado)
}
\arguments{
\item{wynik}{Jest wynikiem wylosowanym w rzucie kostkami.}

\item{ilosc_kostek}{Ilosc kostek na ktorej wypadl dany wynik.}

\item{stan_stada_gracza}{Aktualny stan liczbowy zwierzat w stadzie gracza.}

\item{stado}{Zadane stado, czyli maksymalna ilosc zwierzat danego rodzaju w grze. Domyslnie
stan stada opisany w oryginalnej grze.}
}
\value{
Funkcja \code{rozmnoz_stado} zwraca zmieniony - powiekszony o wynik rzutu
stan stada gracza.
}
\description{
Funkcja \code{rozmnoz_stado} przyjmujac jako argumenty wynik rzutu kostek,
ilosc kostek, na jakiej wypadl i przyjety stan stada powieksza ilosc odpowiednich zwierzat
w stadzie gracza o najwieksza mozliwa wartosc dla zadanej ilosci poszczegolnych zwierzat w grze.
}
\author{
Alicja Gosiewska, \email{alicjagosiewska@gmail.com}

Joanna Zbijewska, \email{asia.zbijewska@gmail.com}
}

