% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rysuj_gestosc.R
\name{rysuj_gestosc}
\alias{rysuj_gestosc}
\title{Funkcja tworzaca wykresy gestosci.}
\usage{
rysuj_gestosc(dane)
}
\arguments{
\item{dane}{Data Frame, z których stworzony zostanie wykres.}
}
\value{
Funkcja \code{badaj_gre} zwraca wykres gestosci z pakietu ggplot2.
}
\description{
Funcja \code{rysuj_gestosc} tworzy wykres gestosci na podstawie zadanych danych i innych parametrów.
}
\author{
Alicja Gosiewska, \email{alicjagosiewska@gmail.com}
}

