% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rysuj_histogram.R
\name{rysuj_histogram}
\alias{rysuj_histogram}
\title{Funkcja tworzaca histogram.}
\usage{
rysuj_histogram(dane, kolor, x_max, y_max)
}
\arguments{
\item{dane}{Dane, z których stworzony zostanie histogram.}

\item{kolor}{Kolor wykresu.}

\item{x_max}{Maksymalna wartość na osi x .}

\item{y_max}{Maksymalna wartość na osi y.}
}
\value{
Funkcja \code{badaj_gre} zwraca histogram w postaci wykresu z pakietu ggplot2.
}
\description{
Funcja \code{rysuj_histogram} tworzy hstogram na podstawie zadanych danych i innych parametrow.
}
\author{
Alicja Gosiewska, \email{alicjagosiewska@gmail.com}

Joanna Zbijewska, \email{asia.zbijewska@gmail.com}
}

