% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rysuj_skrzypcowy.R
\name{rysuj_skrzypcowy}
\alias{rysuj_skrzypcowy}
\title{Funkcja tworzaca wykres skrzypcowy.}
\usage{
rysuj_skrzypcowy(dane)
}
\arguments{
\item{dane}{Data Frame, z których stworzony zostanie wykres.}
}
\value{
Funkcja \code{badaj_gre} zwraca wykres skrzypcowy z pakietu ggplot2.
}
\description{
Funcja \code{rysuj_skrzypcowy} tworzy wykres skrzypcowy na podstawie zadanych danych i innych parametrów.
}
\author{
Alicja Gosiewska, \email{alicjagosiewska@gmail.com}
}

