% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rysuj_wizytowke.R
\name{rysuj_wizytowke}
\alias{rysuj_wizytowke}
\title{Rysowanie wizytówki dla strategii}
\usage{
rysuj_wizytowke(strategia, plik = "wizytowka.pdf", nazwa = NULL)
}
\arguments{
\item{strategia}{strategia do zwizualizowania}

\item{plik}{nazwa pliku, no którego zostanie zapisana wizytówka}

\item{nazwa}{opcjonalna nazwa strategii}
}
\description{
Funkcja rysuje wizytówkę do pliku pdf przedstawiającą podstawowe statystyki dotyczące badanej strategii, w odniesieniu do strategii SuperFarmerDA
}
\author{
Michal Kotlewski ,\email{michal.kotlewski@student.uw.edu.pl}

Mateusz Siwiec, \email{ms332521@students.mimuw.edu.pl}
}

