% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rzuc_i_zareaguj.R
\name{rzuc_i_zareaguj}
\alias{rzuc_i_zareaguj}
\title{Wywolanie i reakcja na wynik rzutu kostek.}
\usage{
rzuc_i_zareaguj(stan_stada_gracza, stado = SuperFarmerAGOJZB::param$stado)
}
\arguments{
\item{stan_stada_gracza}{Aktualny stan zwierzat w stadzie gracza.}

\item{stado}{Mozliwa ilosc zwierzat w stadzie glownym. Domyslnie przypisana wartosc
jak z oryginalnej gry planszowej.}
}
\value{
Funkcja \code{rzuc_i_zareaguj} zwraca uaktualniony i zmieniony o odpowiednia
wartosc stan zwierzat w sadzie gracza.
}
\description{
Funkcja \code{rzuc_i_zareaguj} wywoluje rzut kostek i reaguje na jego
wynik, jako argumenty przyjmuje stan stada gracza oraz maksymalny stan stada.
}
\author{
Alicja Gosiewska, \email{alicjagosiewska@gmail.com}

Joanna Zbijewska, \email{asia.zbijewska@gmail.com}
}

