% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spr_popr_strategii.R
\name{spr_popr_strategii}
\alias{spr_popr_strategii}
\title{Sprawdzenie poprawnosci wymiany}
\usage{
spr_popr_strategii(stan_stada_gracza, tymczasowy_stan_stada_gracza)
}
\arguments{
\item{stan_stada_gracza}{Aktualny stan zwierzat w stadzie gracza.}

\item{tymczasowy_stan_stada_gracza}{Stan stada po dokonaniu wymiany przez strategię.}
}
\value{
Wartosc logiczna oznaczajaca czy wymiana byla legalna.
}
\description{
Funkcja sprawdza, czy wymiana dokonana przez strategię była zgodna z zasadami gry.
}
\author{
Alicja Gosiewska, \email{alicjagosiewska@gmail.com}

Joanna Zbijewska, \email{asia.zbijewska@gmail.com}
}

