% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stan_stada_w_krolikach.R
\name{stan_stada_w_krolikach}
\alias{stan_stada_w_krolikach}
\title{Przeliczenie wartosci wszystkich zwierzat w stadzie gracza na kroliki.}
\usage{
stan_stada_w_krolikach(stan_stada_gracza)
}
\arguments{
\item{stan_stada_gracza}{Aktualny stan liczbowy zwierzat w stadzie gracza.}
}
\value{
Zwraca saldo, czyli ilosc krolikow, na ktore moznaby zamienic wszystkie
obecne w stadzie zwierzeta.
}
\description{
Funckja \code{stan_stada_w_krolikach} przelicza wartosci wszystkich
obecnych aktualnie w stadzie gracza zwierzat na kroliki.
}
\author{
Alicja Gosiewska, \email{alicjagosiewska@gmail.com}
}

