% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_AGOJZB_1owca.R
\name{strategia_AGOJZB}
\alias{strategia_AGOJZB}
\title{Strategia gry w Super Farmera dla pojedynczego gracza, ktory chce zawsze posiadacco najmniej 1 owce.}
\usage{
strategia_AGOJZB(stan_stada_gracza, kroliki = 0, owce = 1, duze_psy = 0,
  male_psy = 0, param = SuperFarmerAGOJZB::param)
}
\arguments{
\item{stan_stada_gracza}{Aktualna ilość zwirząt każdego typu w stadzie gracza.}

\item{kroliki}{Minimalna ilosc krolikow, ktora bedzie chcial zachowac gracz. Domyslnie 0.}

\item{owce}{Minimalna ilosc owiec, ktora bedzie chcial zachowac gracz. Domyslnie 1.}

\item{duze_psy}{Minimalna ilosc duzych_psow, ktora bedzie chcial posiadac gracz. Domyslnie 0.}

\item{male_psy}{Minimalna ilosc malych psow, ktora bedzie chcial posiadac gracz. Domyslnie 0.}

\item{param}{Parametry okreslaja niezbedne wartosci zdefiniowane do gry.}
}
\value{
Funkcja \code{strategia_AGOJZB} zwraca zmieniony o wymienione zwierzeta
stan stada gracza.
}
\description{
Funkcja \code{strategia_AGOJZB} jest strategia gry w gre Super Farmer,
w ktorej gracz chce posiadac zawsze w stadzie co najmniej 1 owce,
strategia zaklada wykonanie wymian koniecznych do jak najszybszej wygranej.
}
\author{
Alicja Gosiewska, \email{alicjagosiewska@gmail.com}
}

