% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_AGOJZB_kroliki.R
\name{strategia_AGOJZB_kroliki}
\alias{strategia_AGOJZB_kroliki}
\title{Strategia gry w Super Farmera dla pojedynczego gracza, ktory chce posiadac jak najwiecej krolikow.}
\usage{
strategia_AGOJZB_kroliki(stan_stada_gracza, kroliki = 9, owce = 0,
  duze_psy = 0, male_psy = 2, param = SuperFarmerAGOJZB::param)
}
\arguments{
\item{stan_stada_gracza}{Aktualna ilość zwirząt każdego typu w stadzie gracza.}

\item{kroliki}{Minimalna ilosc krolikow, ktora bedzie chcial zachowac gracz. Domyslnie 9.}

\item{owce}{Minimalna ilosc owiec, ktora bedzie chcial zachowac gracz. Domyslnie 0.}

\item{duze_psy}{Minimalna ilosc duzych_psow, ktora bedzie chcial posiadac gracz. Domyslnie 0.}

\item{male_psy}{Minimalna ilosc malych psow, ktora bedzie chcial posiadac gracz. Domyslnie 2.}

\item{param}{Parametry okreslaja niezbedne wartosci zdefiniowane do gry.}
}
\value{
Funkcja \code{strategia_AGOJZB_kroliki} zwraca zmieniony o wymienione zwierzeta
stan stada gracza.
}
\description{
Funkcja \code{strategia_AGOJZB_kroliki} jest strategia gry w gre Super Farmer,
w ktorej gracz chce posiadac zawsze w stadzie co najmniej 2 male psy i jak najwiecej krolikow,
strategia zaklada wykonanie wymian koniecznych do jak najszybszej wygranej.
}
\author{
Alicja Gosiewska, \email{alicjagosiewska@gmail.com}

Joanna Zbijewska, \email{asia.zbijewska@gmail.com}
}

