% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strategia_oszukancza.R
\name{strategia_oszukancza}
\alias{strategia_oszukancza}
\title{Strategia niezgodna z zasadami gry SuperFarmer.}
\usage{
strategia_oszukancza(stan_stada_gracza)
}
\arguments{
\item{stan_stada_gracza}{Aktualny stan stada gracza.}
}
\value{
Funkcja \code{strategia_oszukancza} zwraca zmieniony o wymienione zwierzeta
stan stada gracza.
}
\description{
Funkcja \code{strategia_oszukancza} implementuje strategie gry \strong{niezgodna
z zasadami gry} SuperFarmer. Powieksza ona stan stada gracza o jedno zwierze kazdego rodzaju.
}
\author{
Alicja Gosiewska, \email{alicjagosiewska@gmail.com}

Joanna Zbijewska, \email{asia.zbijewska@gmail.com}
}

