% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wartosc.R
\name{wartosc}
\alias{wartosc}
\title{Wartosc danego zwierzecia}
\usage{
wartosc(zwierze)
}
\arguments{
\item{zwierze}{Zwierze, dla ktorego ma byc zwrocona wartosc.}
}
\description{
Funkcja wartosc() zwraca wartosc wskazanego zwiarzecia.
}

