% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wywolaj_n_razy.R
\name{wywolaj_n_razy}
\alias{wywolaj_n_razy}
\title{Wywolanie gry okreslona ilosc razy.}
\usage{
wywolaj_n_razy(n, strategia = SuperFarmerAGOJZB::strategia_AGOJZB,
  param = SuperFarmerAGOJZB::param, sprawdzanie = TRUE)
}
\arguments{
\item{n}{Ilosc razow, jaka chcemy powtorzyc gre.}

\item{strategia}{Argument okreslajacy wykorzystywna do gry strategie, ktorego domyslna
wartosc to strategia zawarta w pakiecie.}

\item{param}{Parametry okreslaja niezbedne wartosci zdefiniowane do gry.}

\item{sprawdzanie}{Argument decydujacy, czy sprawdzenie poprawnosci strategi ma byc wykonane.
Domyslnie jest wykonwane.}
}
\value{
Zwraca wektor o ilosci elementow identycznej do ilosci powtorzen gry, ktory
zawiera ilosc ruchow w jakiej dana gra sie skonczyla.
}
\description{
Funkcja \code{wywolaj_n_razy} powtarza gre w Super Farmera zadana ilosc razy.
}
\author{
Alicja Gosiewska, \email{alicjagosiewska@gmail.com}

Joanna Zbijewska, \email{asia.zbijewska@gmail.com}
}

